+++ controllers/contacts/getAllContacts.js
@@ -1,4 +1,4 @@
-const { Contact } = require("../models")
+const { Contact } = require("../../models")
 
 const getAllContacts = async (req, res) => {
     const contacts = await Contact.find({owner: req.user._id});
