+++ controllers/contacts/getContactById.js
@@ -1,6 +1,6 @@
 const { NotFound } = require('http-errors')
 
-const { Contact } = require("../models");
+const { Contact } = require("../../models");
 
 const getContactById = async (req, res) => {
   const { id } = req.params;
