+++ controllers/users/getCurrent.js
@@ -1,0 +1,9 @@
+const getCurrent = async (req, res) => {
+    const { email } = req.user;
+    res.status(200).json({
+    "email": email,
+    "subscription": "starter"
+    })
+}
+
+module.exports = getCurrent;
