+++ controllers/users/index.js
@@ -1,0 +1,11 @@
+const signup = require('./signup');
+const login = require('./login');
+const logout = require('./logout');
+const getCurrent = require('./getCurrent');
+
+module.exports = {
+    signup,
+    login,
+    logout,
+    getCurrent
+}
