+++ package.json
@@ -19,6 +19,7 @@
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.1.4",
     "morgan": "1.10.0",
+    "multer": "^1.4.4",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
